requirements.txt
click==8.0.1
dataclasses==0.8
importlib-metadata==4.6.1
joblib==1.0.1
numpy==1.19.5
Pillow==8.3.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.5.4
threadpoolctl==2.1.0
timm==0.4.12
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.61.2
typing-extensions==3.10.0.0
zipp==3.5.0
